asdf

+++ beebot/config/database_file_manager.py
@@ -29,10 +29,23 @@ class DatabaseFileManager(FileManager):
         self.body = body
         self.files = {}
 
-    def active_memory(self) -> Memory:
+    async def active_memory(self) -> Memory:
         return self.body.current_memory_chain.incomplete_memory
 
-    def read_file(self, file_path: str) -> str:
+    # Can't support sync db access without a lot of headaches
+    def read_file(self, *args, **kwargs):
+        raise NotImplementedError
+
+    def write_file(self, *args, **kwargs):
+        raise NotImplementedError
+
+    def delete_file(self, *args, **kwargs):
+        raise NotImplementedError
+
+    def list_files(self, *args, **kwargs):
+        raise NotImplementedError
+
+    async def aread_file(self, file_path: str) -> str:
         """Reads a file from the virtual file system in RAM.
 
         Args:
@@ -41,16 +54,13 @@ class DatabaseFileManager(FileManager):
         Returns:
             str: The content of the file. If the file does not exist, returns an error message.
         """
-        if (
-            document := DocumentModel.select()
-            .where(DocumentModel.name == file_path)
-            .get()
-        ):
+        document = await DocumentModel.get_or_none(name=file_path)
+        if document:
             return document.content
         else:
             return "Error: File not found"
 
-    def write_file(self, file_path: str, content: str) -> str:
+    async def awrite_file(self, file_path: str, content: str) -> str:
         """Writes to a file in the virtual file system in RAM.
 
         Args:
@@ -60,31 +70,25 @@ class DatabaseFileManager(FileManager):
         Returns:
             str: A success message indicating the file was written.
         """
-        document, _created = DocumentModel.get_or_create(
-            name=file_path, content=content
+        if not await self.active_memory():
+            return ""
+
+        document, _created = await DocumentModel.get_or_create(
+            name=file_path, defaults={"content": content}
         )
 
-        # Get rid of the link between documents with the same filename and the current memory. This looks gross sorry.
-        stale_link = (
-            DocumentMemoryModel.select()
-            .where(
-                (DocumentModel.name == file_path)
-                & (DocumentMemoryModel.memory == self.active_memory().model_object)
-            )
-            .join(DocumentModel)
-            .first()
-        )
+        stale_link = await DocumentMemoryModel.filter(
+            document__name=file_path, memory=(await self.active_memory()).model_object
+        ).first()
+
         if stale_link:
             logger.warning(f"Deleting stale link ID {stale_link.id}")
-            stale_link.delete()
+            await stale_link.delete()
 
-        self.active_memory().add_document(document)
+        await (await self.active_memory()).add_document(document)
         return f"Successfully wrote {len(content.encode('utf-8'))} bytes to {file_path}"
 
-    async def awrite_file(self, file_path: str, content: str) -> str:
-        return self.write_file(file_path, content)
-
-    def delete_file(self, file_path: str) -> str:
+    async def adelete_file(self, file_path: str) -> str:
         """Deletes a file from the virtual file system in RAM.
 
         Args:
@@ -93,26 +97,27 @@ class DatabaseFileManager(FileManager):
         Returns:
             str: A success message indicating the file was deleted. If the file does not exist, returns an error message.
         """
-        document = DocumentModel.get(name=file_path).first
+        if not await self.active_memory():
+            return
+
+        document = await DocumentModel.get_or_none(name=file_path)
         if not document:
             return f"Error: File not found '{file_path}'"
 
-        document_memory = (
-            DocumentMemoryModel.select()
-            .where(DocumentMemoryModel.document == document)
-            .where(DocumentMemoryModel.memory == self.active_memory().model_object)
-            .get()
-        )
+        document_memory = await DocumentMemoryModel.filter(
+            document=document, memory=(await self.active_memory()).model_object
+        ).first()
+
         if document_memory:
-            document.delete()
+            await document.delete()
         else:
-            # idk why this would happen, perhaps the document hadn't been persisted yet? anyways swallow the error
             logger.warning(
-                f"File {file_path} was supposed to be deleted it does not exist"
+                f"File {file_path} was supposed to be deleted but it does not exist"
             )
+
         return f"Successfully deleted file {file_path}."
 
-    def list_files(self, dir_path: str) -> str:
+    async def alist_files(self, dir_path: str) -> str:
         """Lists all files in the specified directory in the virtual file system in RAM.
 
         Args:
@@ -121,7 +126,13 @@ class DatabaseFileManager(FileManager):
         Returns:
             str: A list of all files in the directory. If the directory does not exist, returns an error message.
         """
-        document_memories = self.active_memory().model_object.document_memories
+        if not await self.active_memory():
+            return ""
+
+        document_memories = await DocumentMemoryModel.filter(
+            memory=(await self.active_memory()).model_object
+        ).prefetch_related("document")
+
         file_paths = [dm.document.name for dm in document_memories]
 
         files_in_dir = [
@@ -134,15 +145,17 @@ class DatabaseFileManager(FileManager):
         else:
             return f"Error: No such directory {dir_path}."
 
-    def all_documents(self) -> list[DocumentModel]:
-        document_memories = (
-            DocumentMemoryModel.select()
-            .where(DocumentMemoryModel.memory == self.active_memory().model_object)
-            .join(DocumentModel)
-        )
+    async def all_documents(self) -> list[DocumentModel]:
+        if not await self.active_memory():
+            return []
+        memory = await self.active_memory()
+        document_memories = await DocumentMemoryModel.filter(
+            memory=memory.model_object
+        ).prefetch_related("document")
+
         return [dm.document for dm in document_memories]
 
-    def load_from_directory(self, directory: str = None):
+    async def load_from_directory(self, directory: str = None):
         if not directory:
             directory = self.body.config.workspace_path
 
@@ -150,21 +163,24 @@ class DatabaseFileManager(FileManager):
             abs_path = os.path.abspath(os.path.join(directory, file.replace("/", "_")))
             if not os.path.isdir(abs_path) and file not in IGNORE_FILES:
                 with open(abs_path, "w+") as f:
-                    self.write_file(file, f.read())
+                    await self.awrite_file(file, f.read())
 
-    def flush_to_directory(self, directory: str = None):
+    async def flush_to_directory(self, directory: str = None):
+        if not await self.active_memory():
+            return
+
         if not directory:
             directory = self.body.config.workspace_path
 
-        for document in self.all_documents():
+        for document in await self.all_documents():
             with open(
                 os.path.join(directory, document.name.replace("/", "_")), "w+"
             ) as f:
                 f.write(document.content)
 
-    def document_contents(self) -> str:
+    async def document_contents(self) -> str:
         documents = []
-        for document in self.all_documents():
+        for document in await self.all_documents():
             file_details = f"## Contents of file {document.name}"
             documents.append(f"\n{file_details}\n{document.content}")
